build fix
git-svn-id: http://www.virtualbox.org/svn/vbox/trunk@46133 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Runtime/common/ldr/ldrPE.cpp
@@ -985,7 +985,7 @@ static int rtldrPEEnumSymbolsSlow(PRTLDRMODPE pThis, unsigned fFlags, RTUINTPTR 
                 const char *pszName = NULL;
                 if (uRvaName != UINT32_MAX)
                 {
-                    size_t cbName = 0x1000 - (uRvaName & 0xfff);
+                    uint32_t cbName = 0x1000 - (uRvaName & 0xfff);
                     if (cbName < 10 || cbName > 512)
                         cbName = 128;
                     rc = rtldrPEReadPartByRva(pThis, NULL, uRvaName, cbName, (void const **)&pszName);
